+++ packaging/build-deb.sh
@@ -118,7 +118,7 @@ Version: $VERSION
 Section: utils
 Priority: optional
 Architecture: $ARCH
-Depends: python3 (>= 3.10), python3-pyqt6, python3-pil, libc6, libstdc++6, libgcc-s1 | libgcc1, libudev1
+Depends: python3 (>= 3.10), python3-pyqt6, python3-pil, python3-psutil, libc6, libstdc++6, libgcc-s1 | libgcc1, libudev1
 Recommends: playerctl, ydotool, wireplumber | pulseaudio-utils, xdg-utils, python3-keyring, python3-pyudev
 Installed-Size: $INSTALLED_KB
 Maintainer: ZoutMax <danielhoutmann@hotmail.com>